requirements.txt
# Core ML
ultralytics>=8.0.0
numpy>=1.24.0
pandas>=2.0.0
opencv-python>=4.8.0
scipy>=1.11.0
matplotlib>=3.7.0

# Dataset download
kagglehub>=0.2.0

# API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Frontend
streamlit>=1.28.0
Pillow>=10.0.0

# Blockchain BSV
bsvlib>=0.10.0
requests>=2.31.0

# Utilities
python-dotenv>=1.0.0
